% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\docType{data}
\name{data}
\alias{data}
\title{Study the different optimization functions}
\format{
An object of class \code{numeric} of length 100.
}
\usage{
data
}
\description{
Study the different optimization functions
}
\keyword{datasets}
