% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\name{loss_net}
\alias{loss_net}
\title{Loss function (quadratic) to find the parameter vector (weights and biases) of the neural net. In the loss function we are collecting the values of weights and bias into one vector: theta = c(w1, w2, b)}
\usage{
loss_net(theta, y, X)
}
\description{
Loss function (quadratic) to find the parameter vector (weights and biases) of the neural net. In the loss function we are collecting the values of weights and bias into one vector: theta = c(w1, w2, b)
}
