% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\name{neural_net}
\alias{neural_net}
\title{The neural net function (a user-friendly wrapper for the optimization)}
\usage{
neural_net(y, X, init.val = rep(0, dim(X)[2] + 1))
}
\description{
The neural net function (a user-friendly wrapper for the optimization)
}
