% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\name{opt_mean1}
\alias{opt_mean1}
\title{Optimization function}
\usage{
opt_mean1(x, fn, iter = 100)
}
\arguments{
\item{x_range}{A \code{vector} representing the response or output.}

\item{fun}{A \code{string} of regressors or inputs.}

\item{B}{A \code{numeric} vector used for starting values (by default a vector of zeros).}
}
\value{
A \code{list} minimum mean
}
\description{
Functions implementing the two optimization routines from previous classes (opt_mean1 and opt_mean2). These take a generic loss function "fn" as an argument.
}
\examples{
First optimization function
}
\author{
Aya
}
